+++ packages/web/src/web/index.ts
@@ -8,3 +8,5 @@ export * from "./language";
 export * from "./broadcastChannel";
 export * from "./geolocation";
 export * from "./cssVariables";
+export * from "./worker";
+export { useWorkerFunction, WebWorkerFunctionOptions } from "./workerFunction";